requirements.txt
cloudscraper
pycryptodome
requests
requests-cache
xxhash>=3.5.0